+++ addOns/retest/src/main/java/org/zaproxy/addon/retest/RetestPlanGenerator.java
@@ -29,6 +29,7 @@ import org.zaproxy.addon.automation.AutomationPlan;
 import org.zaproxy.addon.automation.AutomationProgress;
 import org.zaproxy.addon.automation.ContextWrapper;
 import org.zaproxy.addon.automation.jobs.ActiveScanJob;
+import org.zaproxy.addon.automation.jobs.PolicyDefinition;
 import org.zaproxy.addon.automation.jobs.RequestorJob;
 import org.zaproxy.addon.automation.tests.AbstractAutomationTest;
 import org.zaproxy.addon.automation.tests.AutomationAlertTest;
@@ -101,9 +102,9 @@ public class RetestPlanGenerator {
         ActiveScanJob activeScanJob = new ActiveScanJob();
         activeScanJob.getParameters().setContext("Retest Plan");
         activeScanJob.getData().getPolicyDefinition().setDefaultThreshold("Off");
-        List<ActiveScanJob.Rule> rules = new ArrayList<>();
+        List<PolicyDefinition.Rule> rules = new ArrayList<>();
         rules.add(
-                new ActiveScanJob.Rule(
+                new PolicyDefinition.Rule(
                         alertData.getScanRuleId(), alertData.getAlertName(), "Medium", "Medium"));
         activeScanJob.getData().getPolicyDefinition().setRules(rules);
         activeScanJob.addTest(getAlertTest(alertData, activeScanJob));